+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Fixed
 - `Footer` positioned always at the bottom of the page.
+- ImageUrl warning in `SKUSelector` and `ProductImages`.
 
 ## [1.8.2] - 2018-07-18
 ### Changed